requirements.txt
requests==2.29.0
Flask==2.3.1
redis==4.5.4
google-cloud-spanner==3.35.0
kafka-python